% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extractint.R
\name{Extractint}
\alias{Extractint}
\title{Signal matching and alignment}
\usage{
Extractint(Intensity, Intensitys, next_Intensitys, thr = 5 * 10^(-6))
}
\arguments{
\item{Intensity}{The theoretical m/z value of labeled metabolites in the database.}

\item{Intensitys}{The experimentally detected m/z value of ions.}

\item{next_Intensitys}{The intensity corresponding to the experimentally detected m/z.}

\item{thr}{m/z error range.}
}
\value{
A table containing matched intensities with the targeted m/z.
}
\description{
After inputting the average mass spectra information, this step
    enables automatically matching the intensities between the experimentally
    measured m/z and the targeted list with a tolerance of 5 ppm for each sample.
}
\examples{
\donttest{
Extractint(70.21,70.12,1000,5*10^(-6))
}
}
