% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filterint.R
\name{Filterint}
\alias{Filterint}
\title{Selection of ions with potential labeling signals}
\usage{
Filterint(df, Intensity = 10)
}
\arguments{
\item{df}{The updated table generated after running Scanint.}

\item{Intensity}{The minimum intensity ratio indicating isotopic enrichment.}
}
\value{
A table of potential labeled metabolites.
}
\description{
Compare the average ion intensities of M2 and M3 isotopologue
    between labeled and unlabeled groups by ratio analysis.
}
\examples{
Filterint(Target1, Intensity=10)
}
